Moved command implementations to new file
Cleaned up main discord code and abstracted commands to new file.
